refactor(git): clear the React Doctor findings worth clearing (#294)
Ten of the fifteen warnings on the PR were real. They are fixed here:

- **GitApp was one 415-line component.** Split into three hooks that each own
  something nameable: running an action and reporting a refusal
  (`useGitActions`), what the panes are showing (`useGitSelection`), and how the
  app is laid out (`useGitLayout`). The app file is now composition, 290 lines,
  and was heading for the 500-line rule anyway.
- **Three file rows were divs with click handlers**, so nothing but a mouse
  could reach them. Each is now a button, with the row's own actions as
  siblings — a button cannot contain a button.
- **`transition-all` in two places**, which animates layout as well as colour.
- **`getActionFailureTitle` and `diffContextLabel` sat in a component file**,
  which stops Fast Refresh preserving that file's state. Moved to
  `lib/action-copy.ts`.
- **Three stacked non-null assertions in the conflict parser.** The caller has
  already matched the opening marker, so it passes the label down instead of
  the parser re-matching and asserting.

Left alone, with reasons:

- `conflict-run`'s two awaits inside loops are the point — conflicts are
  resolved one at a time so a question blocks one, not the run.
- `ResolveRunPane`'s `.filter().map()` over a handful of entries: one pass
  through a `reduce` would be slower to read and no faster to run.
- `ModeBanner`'s autoFocus is on an input that only exists because the user
  just clicked to create it; the rule is about focus moving on load.
- `DiffChangeset`'s effect already guards every setter behind its cancelled
  flag.
